removed unneeded files, updated imports
