+++ src/index.ts
@@ -3,8 +3,7 @@ import "./css/index.css";
 import Renderer from "./classes/rendering/Renderer";
 
 // COMPS
-import VertexComp from "./comps/VertexComp";
-import Vertex from "./comps/Vertex";
+import VertexComp from "./comps/Vertex";
 import Button from "./comps/Button";
 
 // INIT FUNC
